Move AuthePreference interface and types back to api/types.
